refactor(sankey chord): sankey图transformData与chord共用;chord去掉weight配置
