+++ src/types/common.ts
@@ -21,25 +21,6 @@ export type Datum = Record<string, any>;
 /** 一个数据序列 */
 export type Data = Datum[];
 
-/** 节点 */
-export type Node = {
-  readonly id: number;
-  readonly name: string;
-};
-
-/** 边 */
-export type Link = {
-  readonly source: number;
-  readonly target: number;
-  readonly value: number;
-};
-
-/** 带有节点与边的数据类型，目前用于桑基图、和弦图 */
-export type NodeLinkData = {
-  readonly nodes: Node[];
-  readonly links: Link[];
-};
-
 /** 点位置信息 */
 export type Point = {
   readonly x: number;
